+++ docs/conf.py
@@ -52,6 +52,7 @@ extensions = [
     'sphinx.ext.intersphinx',
     # 'sphinx.ext.viewcode',
     'sphinx_multiversion',
+    'sphinx_design',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -71,7 +72,7 @@ master_doc = 'index'
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language = "en"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.